[NFC] Extract common reshape patterns to dedicated file
