+++ compiler/src/iree/compiler/Codegen/Common/BUILD.bazel
@@ -159,6 +159,7 @@ iree_compiler_cc_library(
         "RematerializeParallelOps.cpp",
         "RemoveSingleIterationLoop.cpp",
         "ReplaceSlowMinMaxOps.cpp",
+        "ReshapePatterns.cpp",
         "ResolveSwizzleHints.cpp",
         "SpecializeExports.cpp",
         "SplitFullPartialTransferPass.cpp",
