+++ compiler/src/iree/compiler/Codegen/Transforms/Transforms.h
@@ -99,14 +99,6 @@ using GetMinMaxExprFn =
 /// single iteration loops based on the range returned ValueBoundsOpInterface.
 void populateRemoveSingleIterationLoopPattern(RewritePatternSet &patterns);
 
-/// Populate patterns that fold tensor.expand/collapse_shape into the source
-/// hal.interface.binding.subspan.
-void populateReshapeToInterfaceTensorPatterns(RewritePatternSet &patterns);
-
-/// Populate patterns that fold tensor.expand/collapse_shape into the memref
-/// of iree_codegen.load_from_buffer or iree_codegen.store_to_buffer ops.
-void populateFoldTensorReshapeIntoBufferPatterns(RewritePatternSet &patterns);
-
 /// Populate patterns that remove dead allocations
 void populateRemoveDeadMemAllocPatterns(RewritePatternSet &patterns);
